Start rewriting the JS part of the app
